gemini-api/pyproject.toml
[project]
name = "gemini-api"
version = "0.1.0"
description = ""
authors = [
    {name = "uuboyscy",email = "aegis12321@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai (>=1.28.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
